requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.1
openai==1.12.0
httpx==0.23.0
pytest==7.4.3
requests==2.31.0
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4 